WIP - renamed .cpp.h -> .hpp, moved to use 'Visitor' / 'Visit' rather than Context / operator ()
