+++ libscript/src/script-bytecode.hpp
@@ -18,8 +18,8 @@
 
 #include "script-private.h"
 
-#include "script-validate.cpp.h"
-#include "script-execute.cpp.h"
+#include "script-validate.hpp"
+#include "script-execute.hpp"
 
 /*
 struct MCScriptBytecodeOp_Noop
@@ -823,15 +823,13 @@ struct MCScriptBytecodeOp_Reset
 
 #define MC_SCRIPT_DISPATCH_BYTECODE_OP(Name) \
 	case kMCScriptBytecodeOp##Name: \
-		if (!ctxt.template operator()<MCScriptBytecodeOp_##Name>()) \
+		if (!visitor.template Visit<MCScriptBytecodeOp_##Name>()) \
 			return false; \
 		break;
 
-template<
-class Context
->
+template<class Visitor>
 inline bool MCScriptBytecodeDispatch(MCScriptBytecodeOp op,
-									 Context& ctxt)
+									 Visitor& visitor)
 {
 	_Pragma("GCC diagnostic push")
 	_Pragma("GCC diagnostic error \"-Wswitch\"")
@@ -856,65 +854,18 @@ inline bool MCScriptBytecodeDispatch(MCScriptBytecodeOp op,
 	return true;
 }
 
-template<
-typename Context
->
-inline bool MCScriptBytecodeForEach(Context& ctxt)
+#undef MC_SCRIPT_DISPATCH_BYTECODE
+
+
+template<class Visitor>
+inline bool MCScriptBytecodeForEach(Visitor& visitor)
 {
 	for(MCScriptBytecodeOp t_op = kMCScriptBytecodeOp__First; t_op <= kMCScriptBytecodeOp__Last; t_op = (MCScriptBytecodeOp)((int)t_op + 1))
 	{
-		if (MCScriptBytecodeDispatch<Context>(t_op,
-											  ctxt))
+		if (MCScriptBytecodeDispatch(t_op,
+									 visitor))
 			return true;
 	}
 	
 	return false;
 }
-
-#if 0
-template<
-typename Context,
-template<typename OpStruct> class Dispatch
->
-inline bool MCScriptBytecodeDispatch(MCScriptBytecodeOp op,
-									 Context& ctxt)
-{
-	_Pragma("GCC diagnostic push")
-	_Pragma("GCC diagnostic error \"-Wswitch\"")
-	switch(op)
-	{
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(Jump)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(JumpIfFalse)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(JumpIfTrue)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(AssignConstant)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(Assign)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(Return)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(Invoke)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(InvokeIndirect)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(Fetch)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(Store)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(AssignList)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(AssignArray)
-		MC_SCRIPT_DISPATCH_BYTECODE_OP(Reset)
-	}
-	_Pragma("GCC diagnostic pop")
-	
-	return true;
-}
-
-template<
-typename Context,
-template<typename OpStruct> class ForEach
->
-inline bool MCScriptBytecodeForEach(Context& ctxt)
-{
-	for(MCScriptBytecodeOp t_op = kMCScriptBytecodeOp__First; t_op <= kMCScriptBytecodeOp__Last; t_op = (MCScriptBytecodeOp)((int)t_op + 1))
-	{
-		if (MCScriptBytecodeDispatch<Context, ForEach>(t_op,
-													   ctxt))
-			return true;
-	}
-	
-	return false;
-}
-#endif
